requirements.txt
azure-search-documents
azure-core
python-dotenv